chore: format integers to hex in user defined button fields
