+++ lib/presentation/screens/user_defined_buttons/models/button_property_input_field/implementations/button_axis_update_period_millis_input_field.dart
@@ -16,6 +16,7 @@ class ButtonAxisUpdatePeriodMillisInputField
                   .setValue<int?, ButtonAxisUpdatePeriodMillisInputField>,
               mapper: IntParseNullable.tryParseNullable,
               initialValue: initialValue?.toString(),
+              formatIntToUint8: false,
               postMapValidators: (context) => [
                 context.notNullValidator(),
                 context.minValueValidator(0),
