+++ lib/presentation/screens/user_defined_buttons/widgets/input_fields/num_value_getter_input_filed_widget.dart
@@ -173,6 +173,7 @@ class _NumValueGetterInputFieldWidgetState
                                       title:
                                           context.l10n.fractionDigitsFieldTitle,
                                       isRequired: false,
+                                      formatIntToUint8: false,
                                       onChanged: (value) {
                                         addProperty(
                                           FractionsDigitsPropertyEntry(value),
